# logs perf