--- a/sdk/log/cqueue.go
+++ b/sdk/log/cqueue.go
@@ -5,26 +5,26 @@ import (
 	"sync/atomic"
 )
 
+// Cyclic array queue
+//
+// Holds a queue of logging records.
+//
+// When the queue becomes full, the oldest records in the queue are
+// overwritten.
 type cqueue struct {
 	sync.Mutex
 
 	dropped atomic.Uint64
 	cap     int
-	// read, write *ring
 
-	// cyclical array cqueue
 	buf  []*Record
 	head int
 	tail int
 }
 
 func newCQueue(size int) *cqueue {
-	// slog.Info("q init", "cap", size)
-	// r := newRing(size)
 	return &cqueue{
 		cap: size,
-		// read:  r,
-		// write: r,
 		buf: make([]*Record, size+1),
 	}
 }
@@ -52,32 +52,29 @@ func (q *cqueue) Full() bool {
 }
 
 func (q *cqueue) incrPtr(ptr int) int {
-	return (ptr + 1) % (q.Cap() + 1)
+	return q.incrPtrN(ptr, 1)
 }
 
 func (q *cqueue) incrPtrN(ptr int, n int) int {
 	return (ptr + n) % (q.Cap() + 1)
 }
 
-// Dropped returns the number of Records dropped during encqueueing since the
+// Dropped returns the number of Records dropped during enqueueing since the
 // last time Dropped was called.
 func (q *cqueue) Dropped() uint64 {
 	return q.dropped.Swap(0)
 }
 
-// Encqueue adds r to the cqueue. The cqueue size, including the addition of r, is
+// Encqueue adds r to the queue. The queue size, including the addition of r, is
 // returned.
 //
-// If encqueueing r will exceed the capacity of q, the oldest Record held in q
+// If enqueueing r will exceed the capacity of q, the oldest Record held in q
 // will be dropped and r retained.
 func (q *cqueue) Enqueue(r *Record) int {
 	q.Lock()
 	defer q.Unlock()
 
-	// slog.Info("q before encqueue", "len", q.Len(), "tail", q.tail, "head", q.head, "b", b)
-
 	if q.Full() {
-		// slog.Info("q full")
 		q.head = q.incrPtr(q.head)
 		q.dropped.Add(1)
 	}
@@ -85,29 +82,20 @@ func (q *cqueue) Enqueue(r *Record) int {
 	q.buf[q.tail] = r
 	q.tail = q.incrPtr(q.tail)
 
-	// slog.Info("q after encqueue", "len", q.Len(), "tail", q.tail, "head", q.head)
-
 	return q.Len()
 }
 
-// TryDecqueue attempts to decqueue up to len(buf) Records. The available Records
+// TryDequeue attempts to dequeue up to len(buf) Records. The available Records
 // will be assigned into buf and passed to write. If write fails, returning
-// false, the Records will not be removed from the cqueue. If write succeeds,
-// returning true, the decqueued Records are removed from the cqueue. The number
-// of Records remaining in the cqueue are returned.
+// false, the Records will not be removed from the queue. If write succeeds,
+// returning true, the dequeued Records are removed from the queue. The number
+// of Records remaining in the queue are returned.
 //
 // When write is called the lock of q is held. The write function must not call
 // other methods of this q that acquire the lock.
 func (q *cqueue) TryDequeue(buf []*Record, write func([]*Record) bool) int {
 	q.Lock()
 	defer q.Unlock()
-
-	// n := min(len(buf), q.Len())
-	// head := q.head
-	// for i := 0; i < n; i++ {
-	// 	buf[i] = q.buf[head]
-	// 	head = q.incrPtr(head)
-	// }
 
 	n, head := q.cp(buf)
 
@@ -149,13 +137,11 @@ func (q *cqueue) cp(buf []*Record) (n, head int) {
 	return n, head
 }
 
-// Flush returns all the Records held in the cqueue and resets it to be
+// Flush returns all the Records held in the queue and resets it to be
 // empty.
 func (q *cqueue) Flush() []*Record {
 	q.Lock()
 	defer q.Unlock()
-
-	// var ret []*Record
 
 	if q.Empty() {
 		return nil
@@ -167,19 +153,4 @@ func (q *cqueue) Flush() []*Record {
 	q.tail = 0
 
 	return ret
-
-	// if q.Empty() {
-	// 	return nil
-	// } else if q.tail > q.head {
-	// 	ret = q.buf[q.head:q.tail]
-	// 	q.head = 0
-	// 	q.tail = 0
-	// } else {
-	// 	ret = make([]*Record, q.Len())
-	// 	initLen := q.Len()
-	// 	for i := 0; i < initLen; i++ {
-	// 		ret[i] = q.buf[q.head]
-	// 		q.head = q.incrPtr(q.head)
-	// 	}
-	// }
 }