… exporters/otlp/otlplog/otlploghttp/client_test.go
… exporters/otlp/otlplog/otlploghttp/exporter_test.go
… exporters/stdout/stdoutlog/exporter_test.go
… sdk/log/batch_test.go
… sdk/log/exporter_test.go
… sdk/log/logtest/example_test.go
… sdk/log/logtest/factory_test.go
… sdk/log/ring_test.go
… sdk/log/simple_test.go
… triggerci