To resolve: https://github.com/open-telemetry/opentelemetry-go/issues/6382

Resolutions steps:

- refactored to use of pointers to log records instead of passing by values (~1.8x improvement)
- reimplement queue (ring based) as cyclic array queue (~16x improvement)

#### Benchmarks (OTLP)

##### Base
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │                 logs.out                 │
         │     sec/op     │     sec/op      vs base                  │
/OTLP-24      43.58µ ± 6%   2977.49µ ± 10%  +6731.84% (p=0.000 n=10)

         │ spanevents.out │               logs.out                │
         │      B/op      │     B/op       vs base                │
/OTLP-24     99.47Ki ± 5%   193.25Ki ± 1%  +94.29% (p=0.000 n=10)

         │ spanevents.out │               logs.out               │
         │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24      710.0 ± 12%   1861.0 ± 2%  +162.11% (p=0.000 n=10)
```
##### Pointer
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │                 logs.out                 │
         │     sec/op     │     sec/op      vs base                  │
/OTLP-24      43.58µ ± 6%   1617.18µ ± 25%  +3610.63% (p=0.000 n=10)

         │ spanevents.out │                logs.out                │
         │      B/op      │      B/op       vs base                │
/OTLP-24     99.47Ki ± 5%   148.22Ki ± 28%  +49.01% (p=0.002 n=10)

         │ spanevents.out │               logs.out               │
         │   allocs/op    │  allocs/op    vs base                │
/OTLP-24      710.0 ± 12%   1196.5 ± 56%  +68.52% (p=0.023 n=10)
```

##### Pointer Array Queue
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │               logs.out                │
         │     sec/op     │    sec/op     vs base                 │
/OTLP-24      42.44µ ± 7%   178.70µ ± 4%  +321.08% (p=0.000 n=10)

         │ spanevents.out │            logs.out            │
         │      B/op      │     B/op      vs base          │
/OTLP-24     99.34Ki ± 3%   98.13Ki ± 3%  ~ (p=0.280 n=10)

         │ spanevents.out │              logs.out               │
         │   allocs/op    │  allocs/op   vs base                │
/OTLP-24       707.0 ± 7%   400.0 ± 11%  -43.42% (p=0.000 n=10)
```

#### Bench (all)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │ spanevents.out │                 logs.out                 │
           │     sec/op     │     sec/op      vs base                  │
/OTLP-24        46.27µ ± 6%   3180.09µ ± 11%  +6772.15% (p=0.000 n=10)
/STDOUT-24      41.97µ ± 2%    638.88µ ±  1%  +1422.32% (p=0.000 n=10)
/NOOP-24       14.268µ ± 2%     4.496µ ±  1%    -68.49% (p=0.000 n=10)
geomean         30.26µ          209.0µ         +590.81%

           │ spanevents.out │                logs.out                │
           │      B/op      │     B/op       vs base                 │
/OTLP-24       100.5Ki ± 1%    191.4Ki ± 2%   +90.41% (p=0.000 n=10)
/STDOUT-24     90.94Ki ± 0%   274.93Ki ± 0%  +202.33% (p=0.000 n=10)
/NOOP-24       29840.0 ± 0%      240.0 ± 0%   -99.20% (p=0.000 n=10)
geomean        64.34Ki         23.10Ki        -64.09%

           │ spanevents.out │               logs.out               │
           │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24         729.5 ± 3%   1842.0 ± 2%  +152.50% (p=0.000 n=10)
/STDOUT-24       644.5 ± 1%   4625.5 ± 0%  +617.69% (p=0.000 n=10)
/NOOP-24       303.000 ± 0%    3.000 ± 0%   -99.01% (p=0.000 n=10)
geomean          522.3         294.6        -43.60%
```

```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │ spanevents.out │               logs.out                │
           │     sec/op     │    sec/op     vs base                 │
/OTLP-24        44.22µ ± 3%   187.59µ ± 4%  +324.19% (p=0.000 n=10)
/STDOUT-24      40.65µ ± 3%   186.79µ ± 4%  +359.53% (p=0.000 n=10)
/NOOP-24       13.825µ ± 3%    4.346µ ± 1%   -68.57% (p=0.000 n=10)
geomean         29.18µ         53.40µ        +82.99%

           │ spanevents.out │               logs.out                │
           │      B/op      │     B/op       vs base                │
/OTLP-24       99.72Ki ± 7%    97.90Ki ± 1%   -1.82% (p=0.000 n=10)
/STDOUT-24     91.00Ki ± 0%   145.70Ki ± 1%  +60.12% (p=0.000 n=10)
/NOOP-24       29840.0 ± 0%      240.0 ± 0%  -99.20% (p=0.000 n=10)
geomean        64.19Ki         14.95Ki       -76.70%

           │ spanevents.out │               logs.out               │
           │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24        714.5 ± 19%    396.5 ± 3%   -44.51% (p=0.000 n=10)
/STDOUT-24      647.5 ±  1%   1581.0 ± 2%  +144.17% (p=0.000 n=10)
/NOOP-24      303.000 ±  0%    3.000 ± 0%   -99.01% (p=0.000 n=10)
geomean         519.5          123.4        -76.24%
```